# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ShopWave E-Commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wishlist:
    get:
      operationId: getWishlist
      tags: [wishlist]
      summary: Get user's wishlist
      responses:
        "200":
          description: Wishlist items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WishlistItem"
    post:
      operationId: addToWishlist
      tags: [wishlist]
      summary: Add product to wishlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WishlistInput"
      responses:
        "200":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WishlistItem"
  /wishlist/{productId}:
    delete:
      operationId: removeFromWishlist
      tags: [wishlist]
      summary: Remove from wishlist
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Removed

  # ─── ORDERS ────────────────────────────────────────────────────────────────
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/sync:
    post:
      operationId: syncUser
      tags: [users]
      summary: Sync user from Clerk (upsert on login)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserSyncInput"
      responses:
        "200":
          description: Synced user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ─── COUPONS ───────────────────────────────────────────────────────────────
...
  /seller/orders/{orderId}/ship:
    post:
      operationId: shipOrder
      tags: [seller]
      summary: Mark order as shipped with tracking info
      parameters:
        - name: orderId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShipInput"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # ─── ADMIN DASHBOARD ───────────────────────────────────────────────────────
...
components:
  schemas:
...
    WishlistItem:
      type: object
      properties:
        id: { type: integer }
        productId: { type: integer }
        productTitle: { type: string }
        productImage: { type: string }
        productPrice: { type: number }
        finalPrice: { type: number }
        discountPercent: { type: number }
        inStock: { type: boolean }
        avgRating: { type: number }
        addedAt: { type: string }
      required: [id, productId, productTitle, productImage, productPrice, finalPrice, discountPercent, inStock, avgRating, addedAt]
    WishlistInput:
      type: object
      properties:
        productId: { type: integer }
      required: [productId]

    # ─── ORDER ────────────────────────────────────────────────────────────────
    OrderItem:
      type: object
      properties:
        id: { type: integer }
        productId: { type: integer }
        productTitle: { type: string }
        productImage: { type: string }
        variant: { type: ["string", "null"] }
        quantity: { type: integer }
        price: { type: number }
      required: [id, productId, productTitle, productImage, quantity, price]
    OrderTimeline:
      type: object
      properties:
        status: { type: string }
        message: { type: string }
        timestamp: { type: string }
      required: [status, message, timestamp]
    Order:
      type: object
      properties:
        id: { type: integer }
        userId: { type: integer }
        items: { type: array, items: { $ref: "#/components/schemas/OrderItem" } }
        status:
          type: string
          enum: [pending, confirmed, processing, shipped, delivered, cancelled, return_requested, returned]
        paymentMethod: { type: string }
        paymentStatus: { type: string }
        subtotal: { type: number }
        discountAmount: { type: number }
        gstAmount: { type: number }
        shippingFee: { type: number }
        total: { type: number }
        couponCode: { type: ["string", "null"] }
        trackingNumber: { type: ["string", "null"] }
        estimatedDelivery: { type: ["string", "null"] }
        timeline: { type: array, items: { $ref: "#/components/schemas/OrderTimeline" } }
        shippingAddress: { $ref: "#/components/schemas/Address" }
        createdAt: { type: string }
      required: [id, userId, items, status, paymentMethod, paymentStatus, subtotal, discountAmount, gstAmount, shippingFee, total, timeline, createdAt]
...
    ShipInput:
      type: object
      properties:
        trackingNumber: { type: string }
        carrier: { type: string }
        estimatedDelivery: { type: string }
      required: [trackingNumber, carrier, estimatedDelivery]

    # ─── REVIEW ───────────────────────────────────────────────────────────────
...
    Address:
      type: object
      properties:
        id: { type: integer }
        userId: { type: integer }
        name: { type: string }
        phone: { type: string }
        line1: { type: string }
        line2: { type: ["string", "null"] }
        city: { type: string }
        state: { type: string }
        pincode: { type: string }
        country: { type: string }
        isDefault: { type: boolean }
      required: [id, userId, name, phone, line1, city, state, pincode, country, isDefault]
...
    User:
      type: object
      properties:
        id: { type: integer }
        clerkId: { type: string }
        email: { type: string }
        name: { type: string }
        avatarUrl: { type: ["string", "null"] }
        role:
          type: string
          enum: [customer, seller, admin]
        isBanned: { type: boolean }
        createdAt: { type: string }
      required: [id, clerkId, email, name, role, isBanned, createdAt]
    UserUpdate:
      type: object
      properties:
        name: { type: string }
        avatarUrl: { type: ["string", "null"] }
    UserSyncInput:
      type: object
      properties:
        clerkId: { type: string }
        email: { type: string }
        name: { type: string }
        avatarUrl: { type: ["string", "null"] }
      required: [clerkId, email, name]
...